[Codegen] NFC: Lift InnerTiledOp lower & drop-unit-dim patterns to Codegen.
Moves the two `iree_codegen.inner_tiled` rewrite patterns and their
populate functions out of `Codegen/Dialect/GPU/Transforms/` into
`Codegen/Dialect/Codegen/Transforms/LowerInnerTiled.cpp`:

  - `LowerInnerTiledPattern` (replaces an iter-bound-empty `inner_tiled`
    with the per-intrinsic ops emitted by its kind's
    `buildUnderlyingOperations`)
  - `DropInnerTiledUnitDimsPattern` (folds unit iteration bounds and
    pulls the corresponding extract/broadcast pairs out of the loop)

Both patterns operate on `IREE::Codegen::InnerTiledOp` and dispatch
through the kind's interface, so there is nothing GPU-specific about
them. Lifting them lets CPU pass authors share the same lowering logic
without a `Codegen/Dialect/GPU/...` include.

Renames the populate functions accordingly:

  - `IREE::GPU::populateIREEGPULowerInnerTiledPatterns`
        -> `IREE::Codegen::populateLowerInnerTiledPatterns`
  - `IREE::GPU::populateIREEGPUDropUnitDimsPatterns`
        -> `IREE::Codegen::populateDropInnerTiledUnitDimsPatterns`

The `kShapeCastToIntrinsic` / `kShapeCastFromIntrinsic` and
`kDropUnitDims` / `kAddUnitDims` HoistableConversionOp tag constants
move with the patterns. The remaining tag pair on the GPU side
(`kUnrollAccDistribute` / `kUnrollAccReassemble`) stays in
`Codegen/Dialect/GPU/Transforms/Transforms.cpp` since it's still only
used by the GPU unroll path.

Updates the existing in-tree callers in `LowerIREEGPUOps.cpp`,
`IREEGPUExtensions.cpp`, and `UnrollToIntrinsics.cpp`, and adjusts
BUILD.bazel/CMakeLists.txt deps so GPU's transform libraries depend on
`IREECodegenTransforms`.

Strict NFC; no functional or behavioral change.


Progress towards #24323
